requirements.txt
aiohttp>=3.8.0
psycopg2-binary>=2.9.0
asyncio
certifi>=2023.7.22